requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.11
djangorestframework==3.15.1
gunicorn==21.2.0
packaging==24.0
postgres==4.0
psycopg==3.1.18
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
